Separate multi-file-related cases from main reducer
This makes it easier to figure out how the code works and more clearly
enforces the separation of some domains.
